Successful run with new python interface

+++ source/pyskimage.cpp
@@ -80,7 +80,7 @@ static PyObject* skimage_deconvolve(PyObject *self, PyObject *args) {
 
   if (!PyArg_ParseTuple(args,"O!", &PyList_Type, &points)) { return NULL; }
 
-  psize = PyList_Size(points);
+  psize = (uint32_t)PyList_Size(points);
   for (uint32_t i=0;i<psize;i++) {
     datum = PyList_GetItem(points,i);
     atoms.push_back(getvec(datum));
